Continua com erros ao inserir linha. Desta vez parece ter a ver com o DocNovo não ter todos os dados que precisa (apesar de ter sido usado PreencheDadosRelacionados antes)

+++ PP_Extens/PP_PPCS/Purchases/ImportaDocs.cs
@@ -456,7 +456,7 @@ namespace PP_PPCS
                             }
 
                             // Verificar se o documento original é com iva incluido
-                            ivaIncluido = _BSO.Consulta(QueriesSQL.GetQuery08(TipoDoc, Serie)).Valor("IvaIncluido");
+                            ivaIncluido = (int)_BSO.Consulta(QueriesSQL.GetQuery08(TipoDoc, Serie)).Valor("IvaIncluido");
                             
                             RSet = _BSO.Consulta(QueriesSQL.GetQuery09(Filial, TipoDoc, Serie, NumDoc.ToString()));
